+++ src/transformers/models/emu3/processing_emu3.py
@@ -17,12 +17,8 @@
 from ...image_utils import ImageInput
 from ...processing_utils import BatchFeature, MultiModalData, ProcessingKwargs, ProcessorMixin, TextKwargs, Unpack
 from ...tokenization_utils_base import PreTokenizedInput, TextInput
-from ...utils import auto_docstring, is_vision_available
-from ...utils.import_utils import requires
-
-
-if is_vision_available():
-    from .image_processing_emu3 import Emu3ImageProcessorKwargs, smart_resize
+from ...utils import auto_docstring
+from .image_processing_emu3 import smart_resize
 
 
 class Emu3TextKwargs(TextKwargs, total=False):
@@ -39,7 +35,6 @@ class Emu3TextKwargs(TextKwargs, total=False):
 
 class Emu3ProcessorKwargs(ProcessingKwargs, total=False):
     text_kwargs: Emu3TextKwargs
-    images_kwargs: Emu3ImageProcessorKwargs
     _defaults = {
         "text_kwargs": {
             "return_for_image_generation": False,
